CitNet/requirements.txt
scipy==0.19.0
pandas==0.22.0
networkx==1.11
jellyfish==0.6.1
matplotlib==2.0.2
numpy==1.12.1
beautifulsoup4==4.6.0
progressbar2==3.38.0
